El build falla porque una página del cliente importa constantes desde un módulo que arrastra next/headers. Voy a separar esas constantes en un módulo sin dependencias de servidor.fixed.
